Refactored online UI framework
Sounds scary, but this is just moving and renaming files.
